+++ src/app/tabs/tabs.page.spec.ts
@@ -1,7 +1,10 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { TranslateService } from '@ngx-translate/core';
+import { provideRouter } from '@angular/router';
 
 import { TabsPage } from './tabs.page';
+import { createTranslateServiceMock } from '../testing/translate-service.mock';
 
 describe('TabsPage', () => {
   let component: TabsPage;
@@ -9,7 +12,14 @@ describe('TabsPage', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [TabsPage],
+      imports: [TabsPage],
+      providers: [
+        provideRouter([]),
+        {
+          provide: TranslateService,
+          useValue: createTranslateServiceMock(),
+        },
+      ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
     }).compileComponents();
   });