+++ src/app/ui/components/get-mobile-app/get-mobile-app.component.spec.ts
@@ -1,8 +1,9 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 
 import { GetMobileAppComponent } from './get-mobile-app.component';
-import { TranslateService } from '@ngx-translate/core';
+import { createTranslateServiceMock } from 'src/app/testing/translate-service.mock';
 
 describe('GetMobileAppComponent', () => {
   let component: GetMobileAppComponent;
@@ -10,11 +11,18 @@ describe('GetMobileAppComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ],
-      imports: [IonicModule.forRoot(), GetMobileAppComponent],
+      declarations: [],
+      imports: [
+        IonicModule.forRoot(),
+        TranslateModule.forRoot(),
+        GetMobileAppComponent,
+      ],
       providers: [
-      { provide: TranslateService, useValue: jasmine.createSpyObj('TranslateService', ['instant', 'get']) }
-    ]
+        {
+          provide: TranslateService,
+          useValue: createTranslateServiceMock(),
+        },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(GetMobileAppComponent);