+++ src/app/ui/components/get-source-code/get-source-code.component.spec.ts
@@ -1,7 +1,9 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 
 import { GetSourceCodeComponent } from './get-source-code.component';
+import { createTranslateServiceMock } from 'src/app/testing/translate-service.mock';
 
 describe('GetSourceCodeComponent', () => {
   let component: GetSourceCodeComponent;
@@ -9,8 +11,18 @@ describe('GetSourceCodeComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ],
-      imports: [IonicModule.forRoot(), GetSourceCodeComponent]
+      declarations: [],
+      imports: [
+        IonicModule.forRoot(),
+        TranslateModule.forRoot(),
+        GetSourceCodeComponent,
+      ],
+      providers: [
+        {
+          provide: TranslateService,
+          useValue: createTranslateServiceMock(),
+        },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(GetSourceCodeComponent);