+++ src/app/ui/components/privacy-policy/privacy-policy.component.spec.ts
@@ -1,7 +1,10 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
+import { of } from 'rxjs';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 
 import { PrivacyPolicyComponent } from './privacy-policy.component';
+import { createTranslateServiceMock } from 'src/app/testing/translate-service.mock';
 
 describe('PrivacyPolicyComponent', () => {
   let component: PrivacyPolicyComponent;
@@ -9,8 +12,18 @@ describe('PrivacyPolicyComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ],
-      imports: [IonicModule.forRoot(), PrivacyPolicyComponent]
+      declarations: [],
+      imports: [
+        IonicModule.forRoot(),
+        TranslateModule.forRoot(),
+        PrivacyPolicyComponent,
+      ],
+      providers: [
+        {
+          provide: TranslateService,
+          useValue: createTranslateServiceMock(),
+        },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(PrivacyPolicyComponent);